Change design for video search
